requirements.txt
transformers>=4.35.0
datasets>=2.14.0
sentence-transformers>=2.2.2
snorkel-metal>=0.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
spacy>=3.7.0
faiss-cpu>=1.7.4
flask>=3.0.0
flask-cors>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
nltk>=3.8.0
wordcloud>=1.9.0
accelerate>=0.24.0
